+++ Blurred-ios/OtherFollowingCell.swift
@@ -10,6 +10,9 @@ import UIKit
 
 class OtherFollowingCell: UITableViewCell {
 
+    @IBOutlet weak var followingUsername: UILabel!
+    @IBOutlet weak var followingName: UILabel!
+    @IBOutlet weak var followingAvatar: UIImageView!
     override func awakeFromNib() {
         super.awakeFromNib()
         // Initialization code