+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Added other profiles.
 * Added follower list to other user's profiles.
 * Made profile divider change to white if user's theme is set to dark mode.
+* Improved error handling (haha again!)
+* Added following list to other user's profiles. (You can view them and stuff)
 ## v0.2 (PRERELEASE)
 * Added more error handling
 * Some more constraints